refactor(agent): extract tool-call summary into toollog package
Move the generic tool-call input summarizer and bulky-field redaction
helper out of package agent into a new, dependency-free
internal/agent/toollog package, exported as toollog.Summarize and
toollog.Redact. It co-locates the pure utility with its only callers
(the Claude, Codex, and Cursor stream parsers) without implying wider
reuse. No behavior change.
